backend/requirements.txt
flask
flask-cors
joblib
scikit-learn
pandas
numpy
xgboost
pymongo
dnspython
requests
gunicorn
fastapi
uvicorn[standard]
python-dotenv